move equation of state outputs to a separate file
