+++ include/aspect/material_model/interface.h
@@ -540,67 +540,6 @@ namespace aspect
     };
 
 
-    /**
-     * A data structure containing output fields that can be filled by the
-     * evaluate() function of an EquationOfState model. It contains those
-     * properties of the MaterialModelOutputs that are connected to the
-     * equation of state (or the thermodynamic properties). Output values
-     * are computed separately for each composition.
-     *
-     * Accordingly, the vectors are the values for each compositional field
-     * at one specific location.
-     */
-    template <int dim>
-    struct EquationOfStateOutputs
-    {
-      /**
-       * Constructor. Initialize the various arrays of this structure with the
-       * given number of compositions.
-       *
-       * @param n_comp The number of vector quantities (in the order in which
-       * the Introspection class reports them) for which input will be
-       * provided, and outputs should be filled.
-       */
-      EquationOfStateOutputs (const unsigned int n_comp);
-
-      /**
-       * Density values for each composition.
-       */
-      std::vector<double> densities;
-
-      /**
-       * Thermal expansion coefficients for each composition. It is defined
-       * as $\alpha = - \frac{1}{\rho} \frac{\partial\rho}{\partial T}$
-       */
-      std::vector<double> thermal_expansion_coefficients;
-
-      /**
-       * Specific heat for each composition.
-       */
-      std::vector<double> specific_heat;
-
-      /**
-       * Compressibility for each composition. The compressibility is defined
-       * as $\kappa = \frac{1}{\rho} \frac{\partial\rho}{\partial p}$.
-       */
-      std::vector<double> compressibilities;
-
-      /**
-       * The product of the change of entropy $\Delta S$ at a phase transition
-       * and the derivative of the phase function $X=X(p,T,\mathfrak c,\mathbf
-       * x)$ with regard to pressure for each composition.
-       */
-      std::vector<double> entropy_derivative_pressure;
-
-      /**
-       * The product of (minus) the change of entropy $-\Delta S$ at a phase
-       * transition and the derivative of the phase function
-       * $X=X(p,T,\mathfrak c,\mathbf x)$ with regard to temperature for
-       * each composition.
-       */
-      std::vector<double> entropy_derivative_temperature;
-    };
-
 
     /**
      * A namespace in which we define how material model outputs should be
